# hand optimised methods