### Background
I am trying to design a simple JIT compiler for doppio. During analysis, I noticed that for [my benchmark](https://gist.github.com/hrj/94784fc6af4b49a112a0), the following methods were the most frequently called:

```
count of calls: method name
------------------------------------
161516: java/lang/Math/min(II)I
147661: java/lang/Object/<init>()V
113636: com/sun/tools/javac/file/ZipFileIndex/get4ByteLittleEndian([BI)I
113633: com/sun/tools/javac/file/ZipFileIndex/access$500([BI)I
90974: com/sun/tools/javac/file/ZipFileIndex/get2ByteLittleEndian([BI)I
90962: com/sun/tools/javac/file/ZipFileIndex/access$400([BI)I
84696: java/lang/String/compareTo(Ljava/lang/String;)I
83489: com/sun/tools/javac/file/ZipFileIndex$Entry/compareTo(Ljava/lang/Object;)I, com/sun/tools/javac/file/ZipFileIndex$Entry/compareTo(Lcom/sun/tools/javac/f
ile/ZipFileIndex$Entry;)I 
```
Implementing the JIT compiler will take some time, but I thought I would hand-compile the above methods for two reasons:
* To see how much performance benefit it fetches.
* Some of these methods are too complex for the first iteration of JIT that I am planning. They have loops and branches that will take a long time to JIT optimise.

### Changes
This PR adds some more "trapped methods" that help improve speed because they are implemented as native JS functions that use NativeStackFrame instead of interpreted methods with BytecodeStackFrame.

### Benchmark result

For the "compile" benchmark:

| Environment | Time before | Time after (improvement percent)  |
| ----------- | ---------- | ---------------------------------- |
| Chromium 48 |   26870    | 23327 (13.2)                       |
| Firefox 44  |   33796    | 29602 (11.6)                       |