@hrj 2016-03-02
If this PR is acceptable, then further work can be:

* to add trapped implementation for all similar methods such as different `Math.min` and `Math.max` variants.
* remove the "simple" trapped methods after JIT is implemented. For example: `com/sun/tools/javac/file/ZipFileIndex/get4ByteLittleEndian`. These don't have branches or loops so JITing them is easy and has predictable tradeoffs (JIT overhead v/s performance improvement).