@jvilk 2016-03-03
I am not willing to add any additional trapped methods, even for speed. It adds additional maintenance burden that we should not attempt to bear. I would rather focus on improving the speed of the bytecode so that these aren't necessary.

If you want to surface an API to register custom trapped methods, I will merge in a PR for that feature, and you can continue to use hand-defined methods in your projects.

I hope your JIT works out! That would certainly be interesting to see. One speed issue with our bytecode interpreter is that we eagerly create stack frames, rather than create the stack frame lazily when it is needed to suspend / create an exception stack trace. Changing how that works is a serious undertaking, but may make it easier to implement a JIT.