@hrj 2016-03-04
I can understand your hesitancy to merge these changes in.

> If you want to surface an API to register custom trapped methods, I will merge in a PR for that feature, and you can continue to use hand-defined methods in your projects.

That's a great idea; thanks! Adding it to my todo list.

> I hope your JIT works out! That would certainly be interesting to see. One speed issue with our bytecode interpreter is that we eagerly create stack frames, rather than create the stack frame lazily when it is needed to suspend / create an exception stack trace. Changing how that works is a serious undertaking, but may make it easier to implement a JIT.

Is lazy creation of stack frames possible with the asynchronous-calling behaviour of doppio? #359 seems a bit relevant.

The JIT I am planning is very simple; in the first cut, it will just coalesce opcodes to avoid hammering the operand-stack, similar to the first point in https://github.com/plasma-umass/doppio/issues/32. Control-flow opcodes (invokes, conditional jumps) will not be included in the JIT trace, atleast in the first cut. I am hoping to write a design document soon, to get your feedback.