@jvilk 2016-03-08
You'd have to change a lot to lazily create stack frames, but it is possible to do. I look forward to seeing your proposal!